Detect flatpak rather than forcing xdg
My earlier patch forcexdg.patch outright deleted the code block
responsible for determining where to place the configuration files.

This one takes a more subtle approach by actually detecting the flatpak
sandbox and leaving the rest of the code intact.

My plan is to submit this patch to upstream dolphin, if they accept it
we'll be able to drop both this patch and nodirtyversion.patch which is
only required as a side-effect of this.